Split behavior association inheritance helpers
